# BFF/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Ogree BFF:",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "paths": {
...
    "/deviceComp/{entity}": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Create new Devices",
        "tags": [
          "Devices"
        ],
        "operationId": "CreateDevices",
        "parameters": [
          {
            "type": "string",
            "description": "database to retrieve devices",
            "name": "entity",
            "in": "path",
            "required": true
          },
          {
            "format": "object",
            "example": "{\"_name\": \"server\", \"group_name\": \"exwipen22\",\"created\": \"2022-07-18\"}",
            "description": "Mandatory: _name, group_name,created.",
            "name": "body",
            "in": "body",
            "required": true
          }
        ]
      }
    },
    "/deviceComp/{entity}/ConnecteTo/{device}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get Devices connected to a device",
        "tags": [
          "Devices"
        ],
        "operationId": "GetDevicesConnectedTo",
        "parameters": [
          {
            "type": "string",
            "description": "database to retrieve devices",
            "name": "entity",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Key of device",
            "name": "device",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Filter devices by key",
            "name": "_key",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Name of device",
            "name": "_name",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Group_name of device",
            "name": "group_name",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Serial number of device",
            "name": "serial",
            "in": "query"
          }
        ]
      }
    },
...
    "/deviceComp/{entity}{device}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Delete Devices by key",
        "tags": [
          "Devices"
        ],
        "operationId": "DeleteDevices",
        "parameters": [
          {
            "type": "string",
            "description": "database to retrieve devices",
            "name": "entity",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "device looking for",
            "name": "device",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/deviceComps/{entity}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get Devices list",
        "tags": [
          "Devices"
        ],
        "operationId": "Devices",
        "parameters": [
          {
            "type": "string",
            "description": "Key of device",
            "name": "_key",
            "in": "query"
          },
          {
            "type": "string",
            "description": "database to retrieve devices",
            "name": "entity",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Name of device",
            "name": "_name",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Group_name of device",
            "name": "group_name",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Serial number of device",
            "name": "serial",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "schema": {
              "items": {
                "$ref": "#/definitions/SuccessResponse"
              }
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "items": {
                "$ref": "#/definitions/ErrorResponse"
              }
            }
          }
        }
      }
    },
...
    "/projects": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "FlutterApp"
        ],
        "summary": "Get a list of projects for the specified user.",
        "operationId": "GetProjects",
        "parameters": [
          {
            "type": "string",
            "default": "user@test.com",
            "description": "Email of the user whose projects are being requested. Example: /api/projects?user=user@test.com",
            "name": "user",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Return all possible projects."
          },
          "400": {
            "description": "Bad Request. Invalid user query param."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Create a new project",
        "produces": [
          "application/json"
        ],
        "tags": [
          "FlutterApp"
        ],
        "operationId": "CreateProjects",
        "parameters": [
          {
            "format": "object",
            "example": "{\"attributes\":[\"domain\"],\"authorLastUpdate\":\"helder\",\"dateRange\":\"01/01/2023-02/02/2023\", \"lastUpdate\":\"02/02/2023\",\"name\":\"test 1\",\"namespace\":\"physical\",\"objects\":[\"siteB\"],\"showAvg\":false, \"showSum\":false,\"permissions\":[\"user@test.com\",\"admin\"]}",
            "description": "Mandatory: name, dateRange, namespace, attributes, objects, permissions, authorLastUpdate, lastUpdate. Optional: showAvg, showSum, isPublic.",
            "name": "body",
            "in": "body",
            "required": true
          }
        ]
      }
    },
    "/projects/{ProjectID}": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "FlutterApp"
        ],
        "summary": "Replace the data of an existing project.",
        "operationId": "UpdateProjects",
        "parameters": [
          {
            "type": "string",
            "default": "1234",
            "description": "ID of the project to update.",
            "name": "ProjectID",
            "in": "path",
            "required": true
          },
          {
            "format": "object",
            "example": "{\"attributes\":[\"domain\"],\"authorLastUpdate\":\"helder\",\"dateRange\":\"01/01/2023-02/02/2023\", \"lastUpdate\":\"02/02/2023\",\"name\":\"test 1\",\"namespace\":\"physical\",\"objects\":[\"siteB\"],\"showAvg\":false, \"showSum\":false,\"permissions\":[\"user@test.com\",\"admin\"]}",
            "description": "Mandatory: name, dateRange, namespace, attributes, objects, permissions, authorLastUpdate, lastUpdate. Optional: showAvg, showSum, isPublic.",
            "name": "body",
            "in": "body",
            "required": true
          }
        ]
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "FlutterApp"
        ],
        "summary": "Delete an existing project.",
        "operationId": "DeleteProjects",
        "parameters": [
          {
            "type": "string",
            "default": "1234",
            "description": "ID of the project to delete.",
            "name": "ProjectID",
            "in": "path",
            "required": true
          }
        ]
      }
    },
...
  },
  "definitions": {
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "description": "Error Response Message\nin: message",
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "ogree-bff/models"
    },
...
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "message": {
          "description": "Error Response Message\nin: message",
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "ogree-bff/models"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}